스토어 결제 ㅇ / 장바구니에서  쿠폰 불러오기

+++ src/main/java/edu/kh/jvj/mypage/model/vo/Coupon.java
@@ -6,7 +6,7 @@ public class Coupon {
 	
 	private int couponNo; // 쿠폰 번호
 	private String couponName; // 쿠폰 이름
-	private int discountPer; // 할인률
+	private Double discountPer; // 할인률
 	
 	private Date provideDate; // 발급일
 	private Date expireDate; // 만료일
@@ -21,7 +21,7 @@ public class Coupon {
 
 
 
-	public Coupon(int couponNo, String couponName, int discountPer, Date provideDate, Date expireDate,
+	public Coupon(int couponNo, String couponName, Double discountPer, Date provideDate, Date expireDate,
 			int couponStatusCode, String couponStatusName, int memberNo) {
 		super();
 		this.couponNo = couponNo;
@@ -60,13 +60,13 @@ public class Coupon {
 
 
 
-	public int getDiscountPer() {
+	public Double getDiscountPer() {
 		return discountPer;
 	}
 
 
 
-	public void setDiscountPer(int discountPer) {
+	public void setDiscountPer(Double discountPer) {
 		this.discountPer = discountPer;
 	}
 
@@ -81,8 +81,8 @@ public class Coupon {
 	public void setProvideDate(Date provideDate) {
 		this.provideDate = provideDate;
 	}
-
-
+	
+	
 
 	public Date getExpireDate() {
 		return expireDate;
